requirements.txt
python-dotenv
ipykernel
pypdf
langchain-text-splitters
sentence-transformers
tqdm
faiss-cpu
pymysql

cryptography
langchain
langchain-community
groq
python-dotenv
streamlit
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
sqlalchemy

